requirements.txt
asgiref==3.7.2
Django==4.2.5
django-filter==23.3
djangorestframework==3.14.0
drf-yasg==1.21.7
importlib-metadata==6.8.0
inflection==0.5.1
packaging==23.1
platformdirs==3.10.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
yapf==0.40.1
zipp==3.17.0
